# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personalized News API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Article {
  id: string;
  title: string;
  /** Title rewritten in the user's preferred style */
  styledTitle: string;
  description: string;
  /** Description rewritten in the user's preferred style */
  styledDescription: string;
  /** @nullable */
  content?: string | null;
  url: string;
  /** @nullable */
  urlToImage?: string | null;
  source: string;
  /** @nullable */
  author?: string | null;
  publishedAt: string;
  topic: string;
  /**
     * A compelling pull quote extracted from the article
     * @nullable
     */
  pullQuote?: string | null;
}

export interface ArticleSection {
  topic: string;
  /** Section headline styled for the topic */
  headline: string;
  articles: Article[];
}

export interface NewsFeed {
  sections: ArticleSection[];
  /** The newspaper name/masthead for this edition */
  masthead: string;
  /** ISO timestamp of last refresh */
  lastRefreshed: string;
  /** ISO timestamp of next scheduled refresh */
  nextRefresh: string;
  style: string;
  /** Edition label e.g. "Morning Edition · June 6, 2026" */
  edition: string;
  totalArticles?: number;
}

export interface FeedStatus {
  /** @nullable */
  lastRefreshed: string | null;
  /** @nullable */
  nextRefresh: string | null;
  isCached: boolean;
  topics: string;
}

export interface TopicSuggestion {
  name: string;
  category: string;
  emoji: string;
}

export interface TrendingTopics {
  topics: TopicSuggestion[];
}

export type GetNewsParams = {
/**
 * Comma-separated list of topics (e.g. "tech,cricket,AI")
 */
topics: string;
/**
 * Reading style preference
 */
style: GetNewsStyle;
/**
 * Force refresh bypassing cache
 */
refresh?: GetNewsRefresh;
};

export type GetNewsStyle = typeof GetNewsStyle[keyof typeof GetNewsStyle];


export const GetNewsStyle = {
  serious: 'serious',
  punchy: 'punchy',
  casual: 'casual',
  genz: 'genz',
} as const;

export type GetNewsRefresh = typeof GetNewsRefresh[keyof typeof GetNewsRefresh];


export const GetNewsRefresh = {
  true: 'true',
  false: 'false',
} as const;

export type GetNewsStatusParams = {
/**
 * Comma-separated topics to check status for
 */
topics: string;
};