# lib/api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personalized News API
 * OpenAPI spec version: 0.1.0
 */

export interface Article {
  id: string;
  title: string;
  /** Title rewritten in the user's preferred style */
  styledTitle: string;
  description: string;
  /** Description rewritten in the user's preferred style */
  styledDescription: string;
  /** @nullable */
  content?: string | null;
  url: string;
  /** @nullable */
  urlToImage?: string | null;
  source: string;
  /** @nullable */
  author?: string | null;
  publishedAt: string;
  topic: string;
  /**
     * A compelling pull quote extracted from the article
     * @nullable
     */
  pullQuote?: string | null;
}