# lib/api-zod/src/generated/types/articleSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personalized News API
 * OpenAPI spec version: 0.1.0
 */
import type { Article } from './article';

export interface ArticleSection {
  topic: string;
  /** Section headline styled for the topic */
  headline: string;
  articles: Article[];
}