# lib/api-zod/src/generated/types/feedStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personalized News API
 * OpenAPI spec version: 0.1.0
 */

export interface FeedStatus {
  /** @nullable */
  lastRefreshed: string | null;
  /** @nullable */
  nextRefresh: string | null;
  isCached: boolean;
  topics: string;
}