# lib/api-zod/src/generated/types/newsFeed.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personalized News API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleSection } from './articleSection';

export interface NewsFeed {
  sections: ArticleSection[];
  /** The newspaper name/masthead for this edition */
  masthead: string;
  /** ISO timestamp of last refresh */
  lastRefreshed: string;
  /** ISO timestamp of next scheduled refresh */
  nextRefresh: string;
  style: string;
  /** Edition label e.g. "Morning Edition · June 6, 2026" */
  edition: string;
  totalArticles?: number;
}