# lib/api-zod/src/generated/types/topicSuggestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personalized News API
 * OpenAPI spec version: 0.1.0
 */

export interface TopicSuggestion {
  name: string;
  category: string;
  emoji: string;
}