# lib/api-zod/src/generated/types/trendingTopics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Personalized News API
 * OpenAPI spec version: 0.1.0
 */
import type { TopicSuggestion } from './topicSuggestion';

export interface TrendingTopics {
  topics: TopicSuggestion[];
}